+++ changelog.md
@@ -13,6 +13,7 @@
 - ADD: Bookmark Folders
 - FIX: out of bounds error middle mouse button at editor tab.
 - ADD: Beautify JSON, XML and SQL Data
+- ADD: Save image from Clipboard into vx_images directory.
 
 ## v0.1.0